[refactor] util package
